\name{neighbors}
\alias{neighbors}
\title{Find neighbors}
\usage{
neighbors(nodes, budget_A)
}
\arguments{
  \item{nodes}{The number of places to divide the budget
  among}

  \item{budget_A}{The allocation of the budget for each
  node}
}
\value{
Returns vector of neighboring budget allocation vectors
}
\description{
Takes the number of nodes and the budget allocation and
detemines neighboring points
}

