\name{pick_point}
\alias{pick_point}
\title{Select Random Point (or budget allocation)}
\usage{
pick_point(nodes, budget)
}
\arguments{
  \item{nodes}{The number of places to divide the budget
  amongst}

  \item{budget}{The amount of money to spend across the
  nodes}
}
\value{
Returns a vector of the allocation of budget across the
given nodes
}
\description{
Takes budget and nodes to generate a random distrubtion of
the budget over the nodes
}

