\name{profit}
\alias{profit}
\title{Creates a profit function}
\usage{
<<<<<<< HEAD
profit(nodes, budget_A)
=======
profit(nodes, budget_A, coeff_vec)
>>>>>>> a821ca52cf2cf842bd56e93d2edd2897fccd0362
}
\arguments{
  \item{nodes}{Number of places to divide budget amongst}

  \item{budget_A}{The division of budget across the nodes}
<<<<<<< HEAD
=======

  \item{coeff_vec}{The vector of random coefficients to be
  multiplied by the profit function}
>>>>>>> a821ca52cf2cf842bd56e93d2edd2897fccd0362
}
\value{
Returns a vector containing the point's linear terms,
squared terms, and cross products
}
\description{
<<<<<<< HEAD
Takes the budget allocation and creates complex landscape
for that particular point
=======
Takes the number of nodes, budget allocation, and randomd
coefficient vector to determine a profit value for that
particular point
>>>>>>> a821ca52cf2cf842bd56e93d2edd2897fccd0362
}

