\name{profit}
\alias{profit}
\title{Creates a profit function}
\usage{
profit(nodes, budget_A)
}
\arguments{
  \item{nodes}{Number of places to divide budget amongst}

  \item{budget_A}{The division of budget across the nodes}
}
\value{
Returns a vector containing the point's linear terms,
squared terms, and cross products
}
\description{
Takes the budget allocation and creates complex landscape
for that particular point
}

