\name{profit_coeff}
\alias{profit_coeff}
\title{Generate random coefficients}
\usage{
profit_coeff(nodes)
}
\arguments{
  \item{nodes}{The number of places to divide the budget
  among}
}
\value{
Returns a vector of the random coefficients (same length as
profit function terms)
}
\description{
Generates random coefficients for profit function in
between -1 and 1
}

